8060d9-Advanced-ASCII-Art-Generator/requirements.txt
# Add your Python dependencies here
Pillow>=9.0.0
colorama>=0.4.6
opencv-python>=4.8.0
tqdm>=4.65.0
PyYAML>=6.0
pytest>=7.0.0
numpy>=1.21.0
pytest-mock>=3.6.1